requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==46.0.5
distro==1.9.0
Django==6.0.2
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-debug-toolbar==6.2.0
django-environ==0.13.0
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.16.1
docstring_parser==0.17.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jiter==0.13.0
kombu==5.6.2
langdetect==1.0.9
numpy==2.4.2
packaging==26.0
pgvector==0.4.2
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
pytest-django==4.12.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
redis==7.2.1
regex==2026.2.28
requests==2.32.5
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tiktoken==0.12.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
